requirements.txt
torch
torchvision
ftfy
regex
tqdm
gradio
sentence-transformers
git+https://github.com/openai/CLIP.git
Pillow
matplotlib
numpy
